+++ NEWS.md
@@ -3,3 +3,11 @@
 * Added a `NEWS.md` file to track changes to the package.
 * Added functions to convert bitboard to square representation and vice-versa.
 * Added basic board initialization functions
+
+# kappchess 0.1.1
+
+* Added a rook attack set calculator
+* Added a bishop attack set calculator
+* Combined rook and bishop sets to define the Queen attack set calculator
+* print_board function now also prints a single bitboard
+* print_board function is prettier, returning ranks and files